Split Tobira module into three files
The main file has gotten fairly large, and this split somewhat makes
sense.
